Update dbc pools without needing RPC
